+++ src/components/select/common-select-props.ts
@@ -5,7 +5,7 @@
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
  */
-import {Props as EmitterProps} from './emitter';
+import {Props as EmitterProps} from './select-emitter';
 
 /** Common props for all select types (native or menu based) */
 export interface CommonSelectProps extends EmitterProps {