+++ src/components/select/index.ts
@@ -7,6 +7,6 @@
  */
 export {Select, Props as SelectProps} from './select';
 export {OptionLabel, Props as OptionLabelProps} from './option-label';
-export {Emitter, Props as EmitterProps} from './emitter';
+export {SelectEmitter, Props as EmitterProps} from './select-emitter';
 export {Option, Props as OptionProps} from './option';
 export {useSelectContext} from './select-context';