+++ src/icons/arrow-bottom-thin-icon.tsx
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2011-2019 Genestack Limited
+ * All Rights Reserved
+ * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF GENESTACK LIMITED
+ * The copyright notice above does not evidence any
+ * actual or intended publication of such source code.
+ */
+import * as React from 'react';
+
+import {createIcon} from '../components/icon';
+
+// tslint:disable max-line-length
+
+/**
+ * Keyboard arrow key icon shows user possible transformation direction
+ * after some interaction has applied.
+ */
+export const ArrowBottomThinIcon = createIcon(
+    <svg viewBox="0 0 20 20">
+        <path
+            d="M6.5 8.5L10 12L13.5 8.5"
+            fill="none"
+            className="gs-svg-icon-stroke"
+            strokeWidth="0.6"
+            strokeLinecap="round"
+        />
+    </svg>
+);