+++ src/utils/use-constol-invalidity.ts
@@ -7,18 +7,15 @@
  */
 import * as React from 'react';
 
-/** Subscribe to native control invalidity */
-export function useControlInvalidity(
-    controlRef: React.RefObject<HTMLInputElement | HTMLSelectElement>,
-    invalid?: boolean
-) {
-    // Control is valid by default
+/** Subscribe to native input invalidity */
+export function useInputInvalidity(inputRef: React.RefObject<HTMLInputElement>, invalid?: boolean) {
+    // Input is valid by default
     const [invalidState, setInvalidState] = React.useState(invalid || false);
 
-    // Use native control validation before browser paint
+    // Use native input validation before browser paint
     React.useLayoutEffect(() => {
-        if (controlRef.current) {
-            const nextInvalidState = !controlRef.current.validity.valid;
+        if (inputRef.current) {
+            const nextInvalidState = !inputRef.current.validity.valid;
             if (nextInvalidState !== invalidState) {
                 setInvalidState(nextInvalidState);
             }